+++ sharding-jdbc-core/src/main/java/com/dangdang/ddframe/rdb/sharding/api/strategy/database/DatabaseShardingAlgorithm.java
@@ -17,7 +17,7 @@
 
 package com.dangdang.ddframe.rdb.sharding.api.strategy.database;
 
-import com.dangdang.ddframe.rdb.sharding.api.strategy.common.ShardingAlgorithm;
+import com.dangdang.ddframe.rdb.sharding.router.strategy.ShardingAlgorithm;
 
 /**
  * 分库算法接口.