+++ sharding-jdbc-core/src/main/java/com/dangdang/ddframe/rdb/sharding/router/strategy/ShardingStrategy.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package com.dangdang.ddframe.rdb.sharding.api.strategy.common;
+package com.dangdang.ddframe.rdb.sharding.router.strategy;
 
 import com.dangdang.ddframe.rdb.sharding.api.ShardingValue;
 import com.dangdang.ddframe.rdb.sharding.parser.result.router.SQLStatementType;